Remove `DynamicEntryTransformer`
